fix(associacao): adiciona entidade associacao

+++ src/database/migrations/20201130140737-table-debito.js
@@ -28,7 +28,7 @@ module.exports = {
       status: Sequelize.STRING,
       tipoDoc: Sequelize.STRING,
       contrato: Sequelize.STRING,
-      valor: Sequelize.DECIMAL,
+      valor: Sequelize.DOUBLE,
       vencimento: Sequelize.STRING,
       createdAt: {
         type: Sequelize.DATE,
